pyproject.toml

[project]
name = "MotionScore"
version = "0.0.2"
dependencies = [
  "flask>=3.0.3",
  "numpy>=2.2.1",
  "nibabel>=5.3.2",
  "SimpleITK>=2.4.0",
  "pybids>=0.18.1",
  "click>=8.1.8",
  "matplotlib>=3.10.0",
  "pandas>=2.2.3",
  "pillow>=11.0.0",
  "seaborn>=0.13.2",
  "Werkzeug>=3.1.3"
]
requires-python = ">=3.11"
authors = [
    {name = "Bricout Charles", email = "bricout.charles@outlook.com"}
]
maintainers = [
  {name = "Bricout Charles", email = "bricout.charles@outlook.com"}
]
description = "A web based tool to grade and caracterize MRI T1w motion easily"
readme = "README.md"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.pytest.ini_options]
pythonpath = "."
addopts = [
    "--import-mode=importlib",
]

[tool.coverage.run]
omit = [".conda/*"]

[tool.pylint.main]
init-hook = "import sys; sys.path.append('.')"
ignore-paths = [".conda","tests/data"]

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    ".conda",
    "data"
]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "B", "SIM","I"]

[tool.mypy]
exclude = [".conda"]